[Feat] [GGFE-188] submitRankHandler 기존 코드로 복구
