+++ components/modal/afterGame/Buttons.tsx
@@ -2,8 +2,6 @@ import {
   ModalButtonContainer,
   ModalButton,
 } from 'components/modal/ModalButton';
-// FIXME - 동작 확인되면 주석 지울 것 (파일도 정리하기)
-// import styles from 'styles/modal/afterGame/AfterGameModal.module.scss';
 
 interface ButtonsProps {
   onCheck: boolean;
@@ -36,9 +34,6 @@ export function Buttons({
             onClick={onSubmit}
             isLoading={isLoading}
           />
-          {/* FIXME - 동작 확인되면 주석 지울 것 */}
-          {/* <Button style={styles.negative} value='다시입력' onClick={onReset} />
-          <Button style={styles.positive} value='제출하기' onClick={onSubmit} /> */}
         </>
       ) : (
         <ModalButton
@@ -47,8 +42,6 @@ export function Buttons({
           onClick={onEnter}
           isLoading={false}
         />
-        // FIXME - 동작 확인되면 주석 지울 것
-        // <Button style={styles.positive} value='확 인' onClick={onEnter} />
       )}
     </ModalButtonContainer>
   );
