+++ hooks/modal/aftergame/useSubmitModal.ts
@@ -1,9 +1,8 @@
-// import axios from 'axios';
 import { useSetRecoilState } from 'recoil';
 import { AfterGame, TeamScore } from 'types/scoreTypes';
-import { instance } from 'utils/axios';
+import { instance, isAxiosError } from 'utils/axios';
 import { errorState } from 'utils/recoil/error';
-// import { reloadMatchState } from 'utils/recoil/match';
+import { reloadMatchState } from 'utils/recoil/match';
 import { modalState } from 'utils/recoil/modal';
 
 type rankRequest = {
@@ -18,59 +17,51 @@ type normalRequest = {
   gameId: number;
 };
 
-// FIXME - 동작 확인되면 주석 지울 것
-// type responseTypes = Record<'SUCCESS' | 'DUPLICATED', string>;
+const rankResponse: Record<'SUCCESS' | 'DUPLICATED', string> = {
+  SUCCESS: '결과 입력이 완료되었습니다.',
+  DUPLICATED: '상대가 이미 점수를 입력했습니다.',
+};
 
-// type errorResponse = {
-//   status: number;
-//   message: string;
-//   code: string;
-// };
+const errorCode = ['GM202', 'GM204'] as const;
+
+type errorCodeType = (typeof errorCode)[number];
+
+type errorResponse = {
+  status: number;
+  message: string;
+  code: errorCodeType;
+};
 
 const useSubmitModal = (currentGame: AfterGame) => {
   const setError = useSetRecoilState(errorState);
   const setModal = useSetRecoilState(modalState);
-  // const setReloadMatch = useSetRecoilState(reloadMatchState);
+  const setReloadMatch = useSetRecoilState(reloadMatchState);
   const { gameId, matchTeamsInfo, mode } = currentGame;
   const { myTeam, enemyTeam } = matchTeamsInfo;
 
-  // const rankResponse: responseTypes = {
-  //   SUCCESS: '결과 입력이 완료되었습니다.',
-  //   DUPLICATED: '상대가 이미 점수를 입력했습니다.',
-  // };
-
   const submitRankHandler = async (result: TeamScore) => {
-    const requestBody: rankRequest = {
-      gameId: gameId,
-      myTeamId: myTeam.teamId,
-      myTeamScore: result.myTeamScore,
-      enemyTeamId: enemyTeam.teamId,
-      enemyTeamScore: result.enemyTeamScore,
-    };
-    await instance.post(`/pingpong/games/rank`, requestBody);
-    // FIXME - 동작 확인되면 주석 지울 것
-    // try {
-    //   const requestBody: rankRequest = {
-    //     gameId: gameId,
-    //     myTeamId: myTeam.teamId,
-    //     myTeamScore: result.myTeamScore,
-    //     enemyTeamId: enemyTeam.teamId,
-    //     enemyTeamScore: result.enemyTeamScore,
-    //   };
-    //   await instance.post(`/pingpong/games/rank`, requestBody);
-    //   alert(rankResponse['SUCCESS']);
-    // } catch (e) {
-    //   if (axios.isAxiosError(e) && e.response) {
-    //     const { code } = e.response.data as unknown as errorResponse;
-    //     if (code == 'GM202' || code == 'GM204') {
-    //       alert(rankResponse['DUPLICATED']);
-    //       setReloadMatch(true); // 현재 유저 event 상태 재확인
-    //     }
-    //   } else {
-    //     setError('JH04');
-    //     return;
-    //   }
-    // }
+    try {
+      const requestBody: rankRequest = {
+        gameId: gameId,
+        myTeamId: myTeam.teamId,
+        myTeamScore: result.myTeamScore,
+        enemyTeamId: enemyTeam.teamId,
+        enemyTeamScore: result.enemyTeamScore,
+      };
+      await instance.post(`/pingpong/games/rank`, requestBody);
+      alert(rankResponse['SUCCESS']);
+    } catch (e) {
+      if (isAxiosError<errorResponse>(e) && e.response) {
+        const { code } = e.response.data;
+        if (code == 'GM202' || code == 'GM204') {
+          alert(rankResponse['DUPLICATED']);
+          setReloadMatch(true); // 현재 유저 event 상태 재확인
+        }
+      } else {
+        setError('JH04');
+        return;
+      }
+    }
     // openStatChangeModal();
   };
